requirements.txt
djangorestframework>=3.2.4
